requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]==3.12
certifi
scikit-learn
scipy
pyyaml
matplotlib
flask
mlflow


## -e .  # enable and execute it all the end of our project  